Add useElementSize hook and refactor components
